requirements.txt
# Telegram Bot Framework
python-telegram-bot==21.0

# Database
supabase==2.10.0

# Security & Encryption
cryptography==43.0.0
argon2-cffi==23.1.0

# AI Integration
google-generativeai==0.8.3

# Environment Management
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0
